add ShtWalker class and add_filter to add ShtWalker as default walker for wp_nav_menu
